Modified: extract autocomplete from basic/.py files and put it into info.py
