+++ finance/statistics/basic/derivative.py
@@ -1,14 +1,13 @@
 # -*- coding: utf-8 -*-
-import requests
-from bs4 import BeautifulSoup as bs
 from finance.statistics.basic.info import Info
+
 
 class Derivative(Info):
     def __init__(self, code, start, end, day, item, code_to_function, kwargs):
         super().__init__(start, end, day)
         self.function = code_to_function[code]
         if code in ['15002']:
-            self.data_cd, self.data_nm, self.data_tp = self.autocomplete(item)
+            self.data_nm, self.data_cd, self.data_tp = self.autocomplete(item, 'derivative')
         self.item = item
         self.inquiry = kwargs.get('inquiry', None)
         self.trade_index = kwargs.get('trade_index', None)
@@ -17,20 +16,6 @@ class Derivative(Info):
         self.right_type = kwargs.get('right_type', None)
         self.detail = kwargs.get('detail', None)
         self.search_type = kwargs.get('search_type', None)
-
-    def autocomplete(self, item):
-        if item is None:
-            return None, None, None
-        auto_complete_url = 'http://data.krx.co.kr/comm/finder/autocomplete.jspx?contextName=finder_drvprodisu&' \
-                            f'value={item}&viewCount=5&bldPath=%2Fdbms%2Fcomm%2Ffinder%2Ffinder_drvprodisu_autocomplete'
-        response = requests.get(auto_complete_url.format(item=item))
-        soup = bs(response.content, 'html.parser').li
-
-        if soup is None:
-            raise ValueError(f'{item} is Wrong name as a item')
-
-        print(soup.attrs['data-nm'])
-        return soup.attrs['data-cd'], soup.attrs['data-nm'], soup.attrs['data-tp']
 
 
 class ItemPrice(Derivative):
