+++ test.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 import finance
 
-data = finance.per()
+data = finance.data_reader('13302', start=20210401, end=20210406, item_code='52F901')
+
 print(data)
-data = finance.per('LG상사')
-print(data)
-data = finance.per('001120')
-print(data)
+
